reimplement reset calendar dialog
